# swagger.yaml
swagger: "2.0"
info:
  description: >
    <h2>Authentication</h2>
    This server uses auth0.com for authentication.

    <h2>Web Sockets</h2>
    This server uses [socket.io](https://socket.io/) to communicate in real-time between the client and the server when performing lengthy requests, such as in the export process.

    Each interaction is a type of "message" sent to the client from the server. The message is a JSON object that has the following properties:

    **packageId**: The id of the package in which the status message is intended for; in case the same client has requested multiple exports simultaneously - which is not advised.

    **status**: The status of the export (ex: "progress", "error", "complete")

    **message**: A message associated with the status that should be displayed to the user or logged
  version: "1.0.0"
  title: "Trfiolia-on-FHIR"
  contact:
    email: "sean.mcilvenna@lantanagroup.com"
...
host: "##host##"
basePath: "/api"
schemes:
  - "##scheme##"
securityDefinitions:
  auth0:
    type: oauth2
    flow: authorizationCode
    authorizationUrl: https://trifolia.auth0.com/authorize
    tokenUrl: https://trifolia.auth0.com/oauth/token
    scopes: {}
parameters:
  fhirserver:
    name: "fhirserver"
    in: header
    description: "The ID of the configured FHIR server to use for the request. Defaults to the first server in the list of fhir servers as returned by the /config endpoint."
    required: false
    schema:
      type: string
components:
  responses:
    Unauthorized:
      description: The request is not authorized

##swagger-definitions.yaml##
paths:
...
  /import:
    post:
      tags:
        - Import
      security:
        - auth0: []
      parameters:
        - $ref: '#/parameters/fhirserver'
...
  /fhir:
    get:
      summary: "Proxy to underlying FHIR server. All query parameters, path, etc. are redirected to FHIR server specified by the 'fhirserver' header."
      tags:
        - FHIR Proxy
      security:
        - auth0: []
      parameters:
        - $ref: '#/parameters/fhirserver'
...
  /auditEvent:
    get:
      summary: "Retrieves all audit events from the FHIR server"
      tags:
        - Audit
      security:
        - auth0: []
      parameters:
        - $ref: '#/parameters/fhirserver'
      responses:
        '200':
          description: "Success"
          type: array
          items:
            $ref: '#/definitions/AuditEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      summary: "Creates a new audit event"
      tags:
        - Audit
      security:
        - auth0: []
      parameters:
        - $ref: '#/parameters/fhirserver'
      responses:
        '200':
          description: "Success"
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /binary:
    get:
      summary: "Retrieves Binary resources from the FHIR server. Supports any query parameters supported by the FHIR server. See http://build.fhir.org/binary.html#search for possible query parameters."
      descriptions:
      tags:
        - Binary
      security:
        - auth0: []
      responses:
        '200':
          description: "Success"
          type: array
          items:
            $ref: '#/definitions/Binary'
        '401':
          $ref: '#/components/responses/Unauthorized'
      parameters:
        - $ref: '#/parameters/fhirserver'
    post:
      summary: "Creates a new Binary resource on the FHIR server"
      tags:
        - Binary
      security:
        - auth0: []
      parameters:
        - $ref: '#/parameters/fhirserver'
        - in: body
          required: true
          schema:
            $ref: '#/definitions/Binary'
      responses:
        '200':
          description: "Success. Binary resource is returned with the generated ID."
          schema:
            $ref: '#/definitions/Binary'
        '401':
          $ref: '#/components/responses/Unauthorized'
...